Adding new header image and tweaking cards.
