+++ client/my-sites/marketing/business-tools/header.tsx
@@ -8,12 +8,13 @@ import { useTranslate } from 'i18n-calypso';
  * Internal dependencies
  */
 import { Card } from '@automattic/components';
+import headerImage from 'calypso/assets/images/illustrations/business-tools.png';
 
 interface Props {
 	handleButtonClick: () => void;
 }
 
-const MarketingBusinessToolsHeader: FunctionComponent< Props > = ( { handleButtonClick } ) => {
+const MarketingBusinessToolsHeader: FunctionComponent< Props > = () => {
 	const translate = useTranslate();
 
 	return (
@@ -21,7 +22,7 @@ const MarketingBusinessToolsHeader: FunctionComponent< Props > = ( { handleButto
 			<div className="business-tools__header-image-wrapper">
 				<img
 					className="business-tools__header-image"
-					src="/calypso/images/illustrations/illustration-404.svg"
+					src={ headerImage }
 					alt="Your site with Marketing Tools"
 				/>
 			</div>
@@ -33,7 +34,7 @@ const MarketingBusinessToolsHeader: FunctionComponent< Props > = ( { handleButto
 
 				<h2 className="business-tools__header-description">
 					{ translate(
-						'Improve the efficiency and success of your business with the help other these other products and services. Here are some recommended tools for you.'
+						'Improve the efficiency and success of your business with the help of these other products and services. Here are some recommended tools for you.'
 					) }
 				</h2>
 			</div>
